responsiveness aded to the site for better view on mobile screens, Student dashboard and departments logos updated, faculties dashboard has also been updated with minor updates

+++ admin/delete_user.php
@@ -1,7 +1,7 @@
 <?php
 session_start();
 require_once "../includes/db.php";
-// if ($_SESSION['role'] != 'admin') die("Access denied");
+if ($_SESSION['role'] != 'admin') die("Access denied");
 
 $id = $_GET['id'];
 mysqli_query($conn, "DELETE FROM users WHERE id=$id");
